move device_matrix_data kernels to base
