+++ core/base/device_matrix_data_kernels.hpp
@@ -30,8 +30,8 @@ THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ******************************<GINKGO LICENSE>*******************************/
 
-#ifndef GKO_CORE_COMPONENTS_DEVICE_MATRIX_DATA_KERNELS_HPP_
-#define GKO_CORE_COMPONENTS_DEVICE_MATRIX_DATA_KERNELS_HPP_
+#ifndef GKO_CORE_BASE_DEVICE_MATRIX_DATA_KERNELS_HPP_
+#define GKO_CORE_BASE_DEVICE_MATRIX_DATA_KERNELS_HPP_
 
 
 #include <memory>
@@ -93,4 +93,4 @@ GKO_DECLARE_FOR_ALL_EXECUTOR_NAMESPACES(components,
 }  // namespace gko
 
 
-#endif  // GKO_CORE_COMPONENTS_DEVICE_MATRIX_DATA_KERNELS_HPP_
+#endif  // GKO_CORE_BASE_DEVICE_MATRIX_DATA_KERNELS_HPP_
