+++ core/matrix/sparsity_csr.cpp
@@ -42,7 +42,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/matrix/dense.hpp>
 
 
-#include "core/components/device_matrix_data_kernels.hpp"
+#include "core/base/device_matrix_data_kernels.hpp"
 #include "core/components/format_conversion_kernels.hpp"
 #include "core/matrix/sparsity_csr_kernels.hpp"
