+++ reference/matrix/fbcsr_kernels.cpp
@@ -48,9 +48,9 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "accessor/block_col_major.hpp"
 #include "core/base/allocator.hpp"
 #include "core/base/block_sizes.hpp"
+#include "core/base/device_matrix_data_kernels.hpp"
 #include "core/base/iterator_factory.hpp"
 #include "core/base/utils.hpp"
-#include "core/components/device_matrix_data_kernels.hpp"
 #include "core/components/fill_array_kernels.hpp"
 #include "core/components/format_conversion_kernels.hpp"
 #include "core/components/prefix_sum_kernels.hpp"
